+++ runtime/src/main/scala/de/leanovate/jbj/runtime/adapter/LongConverter.scala
@@ -12,7 +12,7 @@ import de.leanovate.jbj.runtime.context.Context
 import de.leanovate.jbj.runtime.types.PParam
 
 object LongConverter extends Converter[Long, IntegerVal] {
-  override def typeName = "integer"
+  override def typeName = "long"
 
   override def missingValue(implicit ctx: Context) = 0L
